requirements.txt
web3==6.11.1
python-dotenv==1.0.0
eth-account==0.9.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
eth-typing==3.5.1